requirements.txt
python-telegram-bot
selenium
sqlalchemy
pymysql